nvidia-nemotron-speech-streaming-en-0.6b/cpu/requirements.txt
huggingface_hub>=0.23.0
nemo_toolkit[asr]>=2.7.1
numpy>=2.3.4
olive-ai>=0.11.0
onnx>=1.20.1
onnxruntime>=1.24.4
sentencepiece>=0.2.1
torch>=2.9.1
